+++ src/rdkafka_conf.c
@@ -1053,7 +1053,11 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
      "authorization server handles. "
      "Only used when `sasl.oauthbearer.method` is set to \"oidc\".",
      _UNSUPPORTED_OIDC},
-    {_RK_GLOBAL, "sasl.oauthbearer.client.secret", _RK_C_STR,
+    {_RK_GLOBAL, "sasl.oauthbearer.client.credentials.client.id", _RK_C_ALIAS,
+     .sdef = "sasl.oauthbearer.client.id"},
+    {_RK_GLOBAL, "sasl.oauthbearer.client.credentials.client.secret",
+     _RK_C_ALIAS, .sdef = "sasl.oauthbearer.client.secret"},
+    {_RK_GLOBAL | _RK_SENSITIVE, "sasl.oauthbearer.client.secret", _RK_C_STR,
      _RK(sasl.oauthbearer.client_secret),
      "Client secret only known to the application and the "
      "authorization server. This should be a sufficiently random string "
@@ -1078,6 +1082,94 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
      "OAuth/OIDC issuer token endpoint HTTP(S) URI used to retrieve token. "
      "Only used when `sasl.oauthbearer.method` is set to \"oidc\".",
      _UNSUPPORTED_OIDC},
+    {
+        _RK_GLOBAL,
+        "sasl.oauthbearer.grant.type",
+        _RK_C_S2I,
+        _RK(sasl.oauthbearer.grant_type),
+        "OAuth grant type to use when communicating with the identity "
+        "provider.",
+        _UNSUPPORTED_OIDC,
+        .vdef = RD_KAFKA_SASL_OAUTHBEARER_GRANT_TYPE_CLIENT_CREDENTIALS,
+        .s2i  = {{RD_KAFKA_SASL_OAUTHBEARER_GRANT_TYPE_CLIENT_CREDENTIALS,
+                  "client_credentials"},
+                 {RD_KAFKA_SASL_OAUTHBEARER_GRANT_TYPE_JWT_BEARER,
+                  "urn:ietf:params:oauth:grant-type:jwt-bearer"}},
+    },
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.algorithm", _RK_C_S2I,
+     _RK(sasl.oauthbearer.assertion.algorithm),
+     "Algorithm the client should use to sign the assertion sent "
+     "to the identity provider and in the OAuth alg header in the JWT "
+     "assertion.",
+     _UNSUPPORTED_OIDC,
+     .vdef = RD_KAFKA_SASL_OAUTHBEARER_ASSERTION_ALGORITHM_RS256,
+     .s2i  = {{RD_KAFKA_SASL_OAUTHBEARER_ASSERTION_ALGORITHM_RS256, "RS256"},
+              {RD_KAFKA_SASL_OAUTHBEARER_ASSERTION_ALGORITHM_ES256, "ES256"}}},
+    {_RK_GLOBAL | _RK_SENSITIVE, "sasl.oauthbearer.assertion.private.key.file",
+     _RK_C_STR, _RK(sasl.oauthbearer.assertion.private_key.file),
+     "Path to client's private key (PEM) used for authentication "
+     "when using the JWT assertion.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL | _RK_SENSITIVE,
+     "sasl.oauthbearer.assertion.private.key.passphrase", _RK_C_STR,
+     _RK(sasl.oauthbearer.assertion.private_key.passphrase),
+     "Private key passphrase for `sasl.oauthbearer.assertion.private.key.file`"
+     " or `sasl.oauthbearer.assertion.private.key.pem`.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL | _RK_SENSITIVE, "sasl.oauthbearer.assertion.private.key.pem",
+     _RK_C_STR, _RK(sasl.oauthbearer.assertion.private_key.pem),
+     "Client's private key (PEM) used for authentication "
+     "when using the JWT assertion.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.file", _RK_C_STR,
+     _RK(sasl.oauthbearer.assertion.file),
+     "Path to the assertion file. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.claim.aud", _RK_C_STR,
+     _RK(sasl.oauthbearer.assertion.claim.audience),
+     "JWT audience claim. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.claim.exp.seconds", _RK_C_INT,
+     _RK(sasl.oauthbearer.assertion.claim.expiration_s),
+     "Assertion expiration time in seconds. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     1, INT_MAX, 300, _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.claim.iss", _RK_C_STR,
+     _RK(sasl.oauthbearer.assertion.claim.issuer),
+     "JWT issuer claim. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.claim.jti.include", _RK_C_BOOL,
+     _RK(sasl.oauthbearer.assertion.claim.jti_include),
+     "JWT ID claim. When set to `true`a random UUID is generated. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     0, 1, 0, _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.claim.nbf.seconds", _RK_C_INT,
+     _RK(sasl.oauthbearer.assertion.claim.not_before_s),
+     "Assertion not before time in seconds. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     0, INT_MAX, 60, _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.claim.sub", _RK_C_STR,
+     _RK(sasl.oauthbearer.assertion.claim.subject),
+     "JWT subject claim. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     _UNSUPPORTED_OIDC},
+    {_RK_GLOBAL, "sasl.oauthbearer.assertion.jwt.template.file", _RK_C_STR,
+     _RK(sasl.oauthbearer.assertion.jwt_template_file),
+     "Path to the JWT template file. "
+     "Only used when `sasl.oauthbearer.method` is set to \"oidc\" and JWT "
+     "assertion is needed.",
+     _UNSUPPORTED_OIDC},
+
 
     /* Plugins */
     {_RK_GLOBAL, "plugin.library.paths", _RK_C_STR, _RK(plugin_paths),
@@ -1416,7 +1508,8 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
     {_RK_GLOBAL | _RK_PRODUCER | _RK_HIGH, "queue.buffering.max.messages",
      _RK_C_INT, _RK(queue_buffering_max_msgs),
      "Maximum number of messages allowed on the producer queue. "
-     "This queue is shared by all topics and partitions. A value of 0 disables "
+     "This queue is shared by all topics and partitions. A value of 0 "
+     "disables "
      "this limit.",
      0, INT_MAX, 100000},
     {_RK_GLOBAL | _RK_PRODUCER | _RK_HIGH, "queue.buffering.max.kbytes",
@@ -1454,7 +1547,8 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
 
     {_RK_GLOBAL | _RK_MED, "retry.backoff.max.ms", _RK_C_INT,
      _RK(retry_backoff_max_ms),
-     "The max backoff time in milliseconds before retrying a protocol request, "
+     "The max backoff time in milliseconds before retrying a protocol "
+     "request, "
      "this is the atmost backoff allowed for exponentially backed off "
      "requests.",
      1, 300 * 1000, 1000},
@@ -1523,7 +1617,8 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
      0, 900000, 10},
     {_RK_GLOBAL, "client.dns.lookup", _RK_C_S2I, _RK(client_dns_lookup),
      "Controls how the client uses DNS lookups. By default, when the lookup "
-     "returns multiple IP addresses for a hostname, they will all be attempted "
+     "returns multiple IP addresses for a hostname, they will all be "
+     "attempted "
      "for connection before the connection is considered failed. This applies "
      "to both bootstrap and advertised servers. If the value is set to "
      "`resolve_canonical_bootstrap_servers_only`, each entry will be resolved "
@@ -1609,7 +1704,8 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
      "`murmur2_random` - Java Producer compatible Murmur2 hash of key "
      "(NULL keys are randomly partitioned. This is functionally equivalent "
      "to the default partitioner in the Java Producer.), "
-     "`fnv1a` - FNV-1a hash of key (NULL keys are mapped to single partition), "
+     "`fnv1a` - FNV-1a hash of key (NULL keys are mapped to single "
+     "partition), "
      "`fnv1a_random` - FNV-1a hash of key (NULL keys are randomly "
      "partitioned).",
      .sdef     = "consistent_random",
@@ -3850,7 +3946,17 @@ const char *rd_kafka_conf_finalize(rd_kafka_type_t cltype,
                                "mutually exclusive";
 
                 if (conf->sasl.oauthbearer.method ==
-                    RD_KAFKA_SASL_OAUTHBEARER_METHOD_OIDC) {
+                        RD_KAFKA_SASL_OAUTHBEARER_METHOD_OIDC &&
+                    !conf->sasl.oauthbearer.token_endpoint_url) {
+                        return "`sasl.oauthbearer.token.endpoint.url` "
+                               "is mandatory when "
+                               "`sasl.oauthbearer.method=oidc` is set";
+                }
+
+                if (conf->sasl.oauthbearer.method ==
+                        RD_KAFKA_SASL_OAUTHBEARER_METHOD_OIDC &&
+                    conf->sasl.oauthbearer.grant_type ==
+                        RD_KAFKA_SASL_OAUTHBEARER_GRANT_TYPE_CLIENT_CREDENTIALS) {
                         if (!conf->sasl.oauthbearer.client_id)
                                 return "`sasl.oauthbearer.client.id` is "
                                        "mandatory when "
@@ -3861,13 +3967,149 @@ const char *rd_kafka_conf_finalize(rd_kafka_type_t cltype,
                                        "mandatory when "
                                        "`sasl.oauthbearer.method=oidc` is set";
                         }
+                }
+                if (conf->sasl.oauthbearer.method ==
+                        RD_KAFKA_SASL_OAUTHBEARER_METHOD_OIDC &&
+                    conf->sasl.oauthbearer.grant_type ==
+                        RD_KAFKA_SASL_OAUTHBEARER_GRANT_TYPE_JWT_BEARER) {
+                        if (conf->sasl.oauthbearer.assertion.file) {
+                                if (conf->sasl.oauthbearer.assertion.private_key
+                                        .file)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.file` cannot both be set";
 
-                        if (!conf->sasl.oauthbearer.token_endpoint_url) {
-                                return "`sasl.oauthbearer.token.endpoint.url` "
-                                       "is mandatory when "
-                                       "`sasl.oauthbearer.method=oidc` is set";
+                                if (conf->sasl.oauthbearer.assertion.private_key
+                                        .pem)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.pem` cannot both be set";
+
+                                if (conf->sasl.oauthbearer.assertion.private_key
+                                        .passphrase)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.passphrase` cannot both be "
+                                               "set";
+
+                                if (conf->sasl.oauthbearer.assertion
+                                        .jwt_template_file)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "jwt.template.file` cannot both "
+                                               "be set";
+
+                                if (conf->sasl.oauthbearer.assertion.claim
+                                        .subject)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "claim.sub` cannot both be set";
+
+                                if (conf->sasl.oauthbearer.assertion.claim
+                                        .audience)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "claim.aud` cannot both be set";
+
+                                if (conf->sasl.oauthbearer.assertion.claim
+                                        .issuer)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "claim.iss` cannot both be set";
+
+                                if (rd_kafka_conf_is_modified(
+                                        conf,
+                                        "sasl.oauthbearer."
+                                        "assertion.claim.jti.include"))
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "claim.jti.include` cannot both "
+                                               "be set";
+
+                                if (rd_kafka_conf_is_modified(
+                                        conf,
+                                        "sasl.oauthbearer."
+                                        "assertion.claim.exp.seconds"))
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "claim.exp.seconds` cannot both "
+                                               "be set";
+
+
+                                if (rd_kafka_conf_is_modified(
+                                        conf,
+                                        "sasl.oauthbearer."
+                                        "assertion.claim.nbf.seconds"))
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "claim.nbf.seconds` cannot both "
+                                               "be set";
+                        } else {
+                                if (conf->sasl.oauthbearer.assertion.private_key
+                                        .file &&
+                                    conf->sasl.oauthbearer.assertion.private_key
+                                        .pem)
+                                        return "Mutually exclusive properties "
+                                               "set. "
+                                               "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.file` and "
+                                               "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.pem` cannot both be set";
+
+                                if (!conf->sasl.oauthbearer.assertion
+                                         .private_key.file &&
+                                    !conf->sasl.oauthbearer.assertion
+                                         .private_key.pem)
+                                        return "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.file` or "
+                                               "`sasl.oauthbearer.assertion."
+                                               "private."
+                                               "key.pem` is mandatory when "
+                                               "`sasl.oauthbearer.grant.type` "
+                                               "is set to "
+                                               "`urn:ietf:params:oauth:grant-"
+                                               "type:jwt-"
+                                               "bearer`";
                         }
                 }
+
+
 
                 /* Enable background thread for the builtin OIDC handler,
                  * unless a refresh callback has been set. */